chore: change scripts to lib
